Move scenario detection to utils
